First pass at QuadLinear interpolation via shader
I believe there could be merit in revisiting the effectively 4D texture
idea from before - it just (ha!) requires a custom shader to do it.
Entities don't render right in sunlight and there are definitely
frequent visual glitches, but I was able to get rid of the banding
problem :)
Overall, this approach has the potential to give us much finer control
over the blending of blocklight and sunlight, if nothing else to closer
mimic the mess that is vanilla lighting.
It's entirely possible some nuance will come along and make this
infeasible, but I'll probably be hanging out on this branch for a bit to
see if this can go anywhere.