Fixed font rendering
Switched FontRenderer to using the Tessellator instead of drawing in
immediate mode, so we get our lighting shader for free. This also
includes a hack to maintain Optifine compatibility (tested with
Optifine_1.7.10_HD_A4.jar), since it sees fit to change an int[] to a
float[] when it replaces FontRenderer, because why not? -.-