Added gamma correction and night vision
I'm not entirely happy with how this commit does night vision (just
changes the floor brightness, so brightness scales from 0.7 to 1.0
instead of 0.05 to 1.0), but the vanilla algorithm looked terrible with
colored lights. Gamma is applied roughly the same as in vanilla, though
colored lights are significantly less pronounced at high gamma,
especially with night vision!