No more flashing sky!
An obnoxiously hacky workaround, but it works and doesn't seem to screw
anything else up. Basically, I wrap the disableLightmap call added in
TransformRendererLivingEntity with a wrapper function that just NOPs
itself and the next enableLightmap call if the shader wasn't enabled in
the first place. I would prefer to take care of this logic in
TransformRendererLivingEntity, but it seems that "just" adding a local
variable to a frame is... complicated. It's doable, and I might work on
that later for fun, but this solution works and puts this in a playable
state :)