Check casting to ChunkCache
This code assumed that every IBlockAccess is either a World or a ChunkCache. StarMiner has a "DummyRotatedBlockAccess" that's a separate implementation of IBlockAccess, and when this code tries to run on it it crashes. This returns a default value of 0 when neither known IBlockAccess type is applicable, presumably causing minor visual weirdness with mods that have custom IBlockAccesses, but not crashing the game.